+++ OtherLanguages/js/src/LercDecode.d.ts
@@ -52,10 +52,32 @@ export interface LercData {
   mask: Uint8Array;
   depthCount: number;
   bandMasks?: Uint8Array[];
+  noDataValues: (number | null)[] | null;
 }
 
+/**
+ * Load the LERC wasm module.
+ *
+ * @param options Use the options to specify a function to locate the wasm file, if not located in the same directory as the script.
+ */
 export function load(options?: { locateFile?: (wasmFileName?: string, scriptDir?: string) => string }): Promise<void>;
+
 export function isLoaded(): boolean;
+
+/**
+ * Decode a LERC byte stream and return an object containing the pixel data.
+ *
+ * @param {object} [options] The decoding options below are optional.
+ * @param {number} [options.inputOffset] The number of bytes to skip in the input byte stream. A valid Lerc file is expected at that position.
+ * @param {number} [options.noDataValue] It is recommended to use the returned mask instead of setting this value.
+ * @param {boolean} [options.returnInterleaved] (ndepth LERC2 only) If true, returned depth values are pixel-interleaved, a.k.a [p1_dep1, p1_dep2, ..., p1_depN, p2_dep1...], default is [p1_dep1, p2_dep1, ..., p1_dep2, p2_dep2...]
+ *
+ */
 export function decode(input: ArrayBuffer | Uint8Array, options?: DecodeOptions): LercData;
+
 export function getBlobInfo(input: ArrayBuffer | Uint8Array, options?: { inputOffset?: number }): LercHeaderInfo;
+
+/**
+ * @deprecated Use getBlobInfo() instead.
+ */
 export function getBandCount(input: ArrayBuffer | Uint8Array, options?: { inputOffset?: number }): number;